Node filled/occluded fix
Reverted back to storing bools on morton-code. And changed debugger to draw nodes, strings and relations in a single loop.
Also added a small calculation for displaying these in a way so that smaller nodes will not be shown until the camera is closer, where the distance gets smaller for each deeper layer so that the screen is not so cluttered.