ACTUAL undo/redo fix
Still haven't really tested all cases. But the redo now works differently in that every undo is checked for how many snapshots were rolled back, where the amount is stored in a list. Then when a redo state is active, we just move the active snapshot index by the batch amount of undos in the last undo operation.