Debug optimization
Now very fast in rendering nodes. It recursively iterates through only the nearby nodes, which skips having to check a lot of tiny nodes since we filter those out by starting from the larger nodes, and stopping the recursion if the larger node is not in range of the camera.