Updater previous-bounds fix
Fixes intersection between prev/curr bounds. Changed the intersection between the two from unrounded to rounded, which is what we want. This gave the wrong remainders and in turn always returned true for checking the remainders for overlap state.

Fixed RecursiveClearUnoccludedChildren which did not update unoccluded node's occluded property correctly.