Refactored FNode
Removed MortonCode on the class itself. Now a FNodePair is used to associate morton-codes with nodes, which was also already accessible from the hashmap iterators.

Also added a SoundPresetID to a this class using the extra space that was freed. Will be used to locate sound-presets which will likely be a thing in the future.